Apply review feedback.
