+++ common/app/common/controllers/PageContent.java
@@ -3,6 +3,16 @@ package common.controllers;
 import io.sphere.sdk.models.Base;
 
 public abstract class PageContent extends Base {
+    private String additionalTitle;
 
-    public abstract String getAdditionalTitle();
+    public PageContent() {
+    }
+
+    public final String getAdditionalTitle() {
+        return additionalTitle;
+    }
+
+    public void setAdditionalTitle(final String additionalTitle) {
+        this.additionalTitle = additionalTitle;
+    }
 }
