+++ purchase/app/purchase/CheckoutThankYouContent.java
@@ -1,19 +1,20 @@
 package purchase;
 
-import common.PageContentBean;
 import common.contexts.UserContext;
+import common.controllers.PageContent;
+import common.controllers.ReverseRouter;
 import common.models.ProductDataConfig;
 import io.sphere.sdk.orders.Order;
 import play.i18n.Messages;
 
-public class CheckoutThankYouContent extends PageContentBean {
+public class CheckoutThankYouContent extends PageContent {
     private CartOrderBean order;
 
     public CheckoutThankYouContent() {
     }
 
-    public CheckoutThankYouContent(final Order order, final UserContext userContext, final ProductDataConfig productDataConfig, final Messages messages) {
-        setOrder(new CartOrderBean(order, userContext, productDataConfig));
+    public CheckoutThankYouContent(final Order order, final UserContext userContext, final ProductDataConfig productDataConfig, final Messages messages, final ReverseRouter reverseRouter) {
+        setOrder(new CartOrderBean(order, userContext, productDataConfig, reverseRouter));
         setAdditionalTitle(messages.at("checkoutThankyouPageTitle"));
     }
